Закончено внедрение БД. Запись в историю из Roll и Custom.

+++ app/src/main/java/com/example/diceroller/data/DatabaseAdapter.java
@@ -9,6 +9,8 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
+// Класс-адаптер для базы данных, предоставляющий
+// абстрактный интерфейс работы с БД
 public class DatabaseAdapter {
 
     private DatabaseHelper dbHelper;
@@ -39,7 +41,7 @@ public class DatabaseAdapter {
         if(cursor.moveToFirst()){
             do{
                 int id = cursor.getInt(cursor.getColumnIndex(DatabaseHelper.COLUMN_ID));
-                Date rollTime = new Date(cursor.getInt(cursor.getColumnIndex(DatabaseHelper.COLUMN_ROLL_TIME)));
+                Date rollTime = new Date(cursor.getLong(cursor.getColumnIndex(DatabaseHelper.COLUMN_ROLL_TIME)));
                 String formulaRaw = cursor.getString(cursor.getColumnIndex(DatabaseHelper.COLUMN_FORMULA_RAW));
                 String formulaProcessed = cursor.getString(cursor.getColumnIndex(DatabaseHelper.COLUMN_FORMULA_PROCESSED));
                 String result = cursor.getString(cursor.getColumnIndex(DatabaseHelper.COLUMN_RESULT));
@@ -61,7 +63,7 @@ public class DatabaseAdapter {
         String query = String.format("SELECT * FROM %s WHERE %s=?",DatabaseHelper.TABLE, DatabaseHelper.COLUMN_ID);
         Cursor cursor = database.rawQuery(query, new String[]{ String.valueOf(id)});
         if(cursor.moveToFirst()){
-            Date rollTime = new Date(cursor.getInt(cursor.getColumnIndex(DatabaseHelper.COLUMN_ROLL_TIME)));
+            Date rollTime = new Date(cursor.getLong(cursor.getColumnIndex(DatabaseHelper.COLUMN_ROLL_TIME)));
             String formulaRaw = cursor.getString(cursor.getColumnIndex(DatabaseHelper.COLUMN_FORMULA_RAW));
             String formulaProcessed = cursor.getString(cursor.getColumnIndex(DatabaseHelper.COLUMN_FORMULA_PROCESSED));
             String result = cursor.getString(cursor.getColumnIndex(DatabaseHelper.COLUMN_RESULT));
